munged: Pick up munge key from CREDENTIALS_DIRECTORY
This adds the ability to automatically pick up the munge key from a systemd
credential if no other key is defined.

A well-known credential name "munge.key" is defined for this purpose and will be
looked up relative to the path in the CREDENTIALS_DIRECTORY environment
variable.

This path sidesteps checks for safety, since credentials are only readable by
the service user and reside in undumpable memory and all safety guarantees come
from the service manager.

[1] https://systemd.io/CREDENTIALS/